requirements.txt
asgiref==3.5.1
colorama==0.4.4
Django==4.0.4
django-admin-rangefilter==0.8.4
face-recognition-models==0.3.0
joblib==1.1.0
numpy==1.22.3
Pillow==9.1.0
scikit-learn==1.0.2
scipy==1.8.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
